Feature/autenticador jwt

+++ src/controllers/produto.ts
@@ -1,6 +1,8 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { Request, Response } from "express";
 import { PrismaClient } from "@prisma/client";
+import verificaPermissao from "../services/verificaPermissao";
+import { Permissoes } from "../data/permissoes";
 
 const prisma = new PrismaClient();
 
@@ -9,6 +11,14 @@ class ProdutoController {
     const nomeProduto = req.body.nomeProduto;
     const unidade = req.body.unidade;
     try {
+      const { authorization } = req.headers;
+      if (!authorization) {
+        return res.status(401)
+      }
+      const permissao = await verificaPermissao.validaPermissao(authorization, Permissoes.PRODUTOS)
+      if (!permissao) {
+        return res.status(401)
+      }
       const produto = await prisma.produto.create({
         data: {
           nomeProduto: nomeProduto,
@@ -24,6 +34,15 @@ class ProdutoController {
 
   async get(req: Request, res: Response) {
     try {
+      const { authorization } = req.headers;
+      if (!authorization) {
+        return res.status(401)
+      }
+      const permissao = await verificaPermissao.validaPermissao(authorization, Permissoes.PRODUTOS)
+      if (!permissao) {
+        return res.status(401)
+      }
+      
       const produtos = await prisma.produto.findMany();
       res.json(produtos);
     } catch (err) {
@@ -34,6 +53,14 @@ class ProdutoController {
 
   async getById(req: Request, res: Response) {
     try {
+      const { authorization } = req.headers;
+      if (!authorization) {
+        return res.status(401)
+      }
+      const permissao = await verificaPermissao.validaPermissao(authorization, Permissoes.PRODUTOS)
+      if (!permissao) {
+        return res.status(401)
+      }
       const id = req.params.id;
       const produto = await prisma.produto.findFirst({
         where: { id: parseInt(id) },
@@ -47,6 +74,14 @@ class ProdutoController {
 
   async getByNome(req: Request, res: Response) {
     try {
+      const { authorization } = req.headers;
+      if (!authorization) {
+        return res.status(401)
+      }
+      const permissao = await verificaPermissao.validaPermissao(authorization, Permissoes.PRODUTOS)
+      if (!permissao) {
+        return res.status(401)
+      }
       const nomeProduto = req.params.nomeProduto;
       const produto = await prisma.produto.findMany({
         where: {
@@ -63,6 +98,14 @@ class ProdutoController {
 
   async put(req: Request, res: Response) {
     try {
+      const { authorization } = req.headers;
+      if (!authorization) {
+        return res.status(401)
+      }
+      const permissao = await verificaPermissao.validaPermissao(authorization, Permissoes.PRODUTOS)
+      if (!permissao) {
+        return res.status(401)
+      }
       const id = req.params.id;
       const nomeProduto = req.body.nomeProduto;
       const unidade = req.body.unidade;
@@ -80,39 +123,45 @@ class ProdutoController {
     }
   }
 
-  async delete(req: Request, res: Response){
+  async delete(req: Request, res: Response) {
     try {
-        const id = parseInt(req.params.id);
-        await prisma.$transaction([
-            prisma.fornecedorProduto.deleteMany({
-              where: {
-                produtoId: id // o ID do registro em ModeloB correspondente que deseja excluir
-              }
-            }),
-            prisma.entregaProduto.deleteMany({
-                where: {
-                    produtoId: id
-                }
-            }),
-            prisma.qualidadeProduto.deleteMany({
-                where: {
-                    produtoId: id
-                }
-            }),
-            prisma.produto.deleteMany({
-                where: {
-                  id: id 
-                }
-              }),
-          ]);
-        res.status(200).json('Produto deletado')
+      const { authorization } = req.headers;
+      if (!authorization) {
+        return res.status(401)
+      }
+      const permissao = await verificaPermissao.validaPermissao(authorization, Permissoes.PRODUTOS)
+      if (!permissao) {
+        return res.status(401)
+      }
+      const id = parseInt(req.params.id);
+      await prisma.$transaction([
+        prisma.fornecedorProduto.deleteMany({
+          where: {
+            produtoId: id, // o ID do registro em ModeloB correspondente que deseja excluir
+          },
+        }),
+        prisma.entregaProduto.deleteMany({
+          where: {
+            produtoId: id,
+          },
+        }),
+        prisma.qualidadeProduto.deleteMany({
+          where: {
+            produtoId: id,
+          },
+        }),
+        prisma.produto.deleteMany({
+          where: {
+            id: id,
+          },
+        }),
+      ]);
+      res.status(200).json("Produto deletado");
     } catch (err) {
-        console.log(err);
-        res.status(400).json(err)
+      console.log(err);
+      res.status(400).json(err);
     }
   }
 }
-
-
 
 export default new ProdutoController();
